+++ CHANGES.md
@@ -1,3 +1,8 @@
+
+* Add `Mtime.Span.{is_shorter,is_longer}` to make duration
+  comparisons more obivous. Thanks to Pau Ruiz Safont for
+  the suggestion and the patch.
+
 v2.0.0 2022-12-02 Zagreb
 ------------------------